+++ src/compiler/dyvil/tools/compiler/ast/expression/LambdaExpr.java
@@ -371,7 +371,7 @@ public final class LambdaExpr implements IValue, IClassCompilable, IDefaultConte
 	}
 
 	@Override
-	public float getTypeMatch(IType type)
+	public int getTypeMatch(IType type)
 	{
 		if (type.getTheClass() == Types.OBJECT_CLASS)
 		{
